requirements.txt
opencv-python
python-dotenv
google-cloud-storage
